My game project using OOP.

+++ WarGame_Project/Character_Settings.cs
@@ -8,6 +8,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using WarGame_Project.Business.Repositories.Concrete;
+using WarGame_Project.Business.Utilities;
 using WarGame_Project.Model.Entities.Concrete;
 
 namespace WarGame_Project
@@ -46,20 +47,31 @@ namespace WarGame_Project
 
         private void btnUpdateCharacter_Click(object sender, EventArgs e)
         {
-            warCharacter.Id = int.Parse(textIdUpdate.Text);
-            warCharacter.Rank = comboBoxRank.Text;
-            warCharacter.Weapon = comboBoxWeapon.Text;  
-            warCharacter.Physic = comboBoxPhysics.Text;
-            warCharacter.Talent = comboBoxTalent.Text;
-            characterRepository.Update(warCharacter);
-            dataGridView1.DataSource = characterRepository.GetAll();
+            if (Utilities.TextControl(grpUpdateCharacter) == false)
+            {
+                warCharacter.Id = int.Parse(textIdUpdate.Text);
+                warCharacter.Rank = comboBoxRank.Text;
+                warCharacter.Weapon = comboBoxWeapon.Text;
+                warCharacter.Physic = comboBoxPhysics.Text;
+                warCharacter.Talent = comboBoxTalent.Text;
+                characterRepository.Update(warCharacter);
+                dataGridView1.DataSource = characterRepository.GetAll();
+
+
+            }
         }
 
         private void btnDeleteCharacter_Click(object sender, EventArgs e)
         {
-            int characterid = int.Parse(textIdDelete.Text); 
-            WarCharacter character = characterRepository.GetById(characterid);  
-            characterRepository.Delete(character);  
+            if (Utilities.TextControl(grpUpdateCharacter) == false)
+            {
+
+                int characterid = int.Parse(textIdDelete.Text);
+                WarCharacter character = characterRepository.GetById(characterid);
+                characterRepository.Delete(character);
+
+            }
+                
         }
 
         private void btnBACK_Click(object sender, EventArgs e)
